cleaned repo of some warnings

+++ autonomous.h
@@ -13,8 +13,8 @@
 #define CLAW_RIGHT 2
 
 void turnTowardsStars(bool isOnRight);
-bool determineTurnNegative(double angle);
-void turn(double angle, int speed);
+bool determineTurnNegative(float angle);
+void turn(float angle, int speed);
 void straighten(int speed);
 void collectStars(bool isOnRight, bool endOnRightSide);
 void startWallDrive();
